+++ src/main/java/com/mindhub/homebanking/models/Loan.java
@@ -18,9 +18,9 @@ public class Loan {
     private Set<ClientLoan> clientLoans = new HashSet<>();
 
     private String name;
-    private Integer maxAmount;
+    private Double maxAmount;
 
-    public Loan(String name, Integer maxAmount, Set<Integer>payments) {
+    public Loan(String name, Double maxAmount, Set<Integer>payments) {
         this.name = name;
         this.maxAmount = maxAmount;
         this.payments=payments;
@@ -53,11 +53,11 @@ public class Loan {
         this.name = name;
     }
 
-    public Integer getMaxAmount() {
+    public Double getMaxAmount() {
         return maxAmount;
     }
 
-    public void setMaxAmount(Integer maxAmount) {
+    public void setMaxAmount(Double maxAmount) {
         this.maxAmount = maxAmount;
     }
